requirements.txt
discord.py==2.1.0
kbbi==0.4.3
